+++ cases/testcases/is_false/green.clj
@@ -1,0 +1,6 @@
+(ns testcases.is-false.green
+  (:require
+   [clojure.test :refer [are deftest is join-fixtures testing use-fixtures]]))
+
+(deftest foo
+  (is false))